agent/requirements.txt
# Pi Control Panel - Agent Dependencies

# Core
pyyaml>=6.0
python-dotenv>=1.0.0

# Unix socket RPC
aiohttp>=3.9.0

# System metrics
psutil>=5.9.0

# Docker integration
docker>=7.0.0

# D-Bus for systemd
dbus-python>=1.3.2
PyGObject>=3.46.0

# MQTT client
paho-mqtt>=2.0.0

# SQLite (built-in, but we need aiosqlite for async)
aiosqlite>=0.19.0

# Logging
structlog>=23.2.0

# Async utilities
asyncio-throttle>=1.0.2

# Type hints
typing-extensions>=4.9.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.0
